TOO MANY CACHES!
After deploying changes to www-ian or www-test, I wasn’t seeing changes reflected. This was partly due to cloudfront caching our index.html file, which I think I’ve fixed now. It’s also due to the service worker running in the react app serving up the old index.html file. I’ve adjusted the logic so that the service worker is only registered when viewing the app on the production site (https://boombox.bingo). Otherwise it checks to see if there’s a service working running and disables it. Since you probably already have a service working caching an old version of the site, you’ll probably need to manually disable it in the “Application” section of the chrome inspector.